Fix subtle implementation bug with maximal set filtering
There was a subtle bug that was resolved involving
traversal over the maximal event set tree of a configuration.
Specifically, it's important to

1. Only pick candidates that satisfy the predicate. This
means that the `bookkeeper` needs to own the reference
to the filter so that is can accurately pick events
correctly

2. At the start, pick the first *viable* candidate, and not simply
the *first* node in the ordering, as the first node may
not satisfy the predicate. Effectively, you pick the
"most recent" node that satisfies the predicate

A possible future addition would be to pre-compute the
predicate for all of the events at the start of the
iterator. This would reduce the number of calls made
to the filter function dramatically, but for now we'll
stick with something simple (the change wouldn't involve
too much intervention if desired)