Add iterator_wrapping convenience container
The iterator_wrapping structure is a light-weight
container whose contents are traversed according
to the behavior of the iterator that it creates
when asked. The wrapping can only construct
a container around an iterator whose termination
condition can be succinctly represented with a
default-constructed argument. It is not always
the case that such a default-constructed iterator
represents the end, so this should be kept in made
(although this is generally the convention).

Creating an instance of a wrapping allows, e.g.,
for the traversal over a collection without explicit
need to construct the iterators/loop manually.
For example, a struct `Iterable` may define a method
`make_complicated_forward_iterator_over_me()`
that can be called and used inside e.g. an auto-based
for-loop. This helps with readability immensely.