Begin filling in computations of ex(C) and en(C)
This commit introduces the first steps towards
computing ex(C) and en(C) using many of the
additions of the previous phases. An important
part of the computation is so far missing, however,
viz. the determination of whether or not an action
`a` is enabled at some point in the past. We'll
need to decipher what's going on in tiny_simgrid
to fully determine what to do here; for now, we
simply ignore adding an event entirely and just
say that `a` was disabled at all points in the past.