Begin tracking latest events for each actor
Computing specialized extensions for each transition
type frequently requires knowing the "latest" event
that some given actor is associated with in some
configuration (denoted `preEvt(C, a)`). We now track
this information as new events are added and pre-compute
the information when constructing a configuration with
a set of events using once again our favorite topological
ordering function.

Subsequent commits will introduce some unit tests to
ensure that the Configuration is tracking the new events
appropriately.