Expand sanity-check tests for CommSend/CommRecv transitions
A very, very basic test was added to ensure that
the most trivial program was treated correctly.
More tests will be added which will make for more
interesting scenarios. The difficulty lies with
creating the unfolding itself: there's no better
way than to do so manually and to follow along
with where UDPOR would take us...