Add partial implementation for MutexWait/MutexTest
The full implementation for computing extension
sets for MutexWait and MutexTest was *almost*
added in this commit. Again, though, we need to
be able to determine dependecies in the context
of mutex transitions which implies we need to
be have more information from the application side.
This work will be added in subsequent commits