Add skeleton of implementation for tree insertion
The `insert` method on a Wakeup tree is an
essential component of the ODPOR algorithm.
Tree insertion guides the ODPOR algorithm
towards the unique, unsearched traces of the
state space. Insertion is based heavily on the
correctness of computing the `~` relation over
a given execution. Subsequent commits will fill
in this method as well as several others that
will be needed to implement that functionality