Add tree pruning/subtree methods to State
ODPOR requires that we can remove a
subtree from a wakeup tree as we continue
the exploration. Further, we must also be
able to create a subtree from a given
wakeup tree. The functionality for those
methods has already been added; this commit
instead introduces the management of wakeup
trees specifically in the context of ODPOR
exploration.