Add ODPOR race detection phase rough-draft
ODPOR is decoposed into two parts: the "passive"
observer phase, where ODPOR selects search points
based on the work it did before to fill in its
wakeup trees, and the active race-detection phase,
where ODPOR looks at all reversible races in an
execution and performs the work to determine if
new traces need be searched. This commit adds
in an outline for the latter phase. Note that
there is still much work to be done with respect
to the race detection -- we still need to implement
the partial execution function for example.

Likewise, handling the issue of multiple execution
possibilities for a given process will be particularly
tricky.