Fix WakeupTreeIterator bug comparing diff iterators
Prior to this commit, the WakeupTreeIterator
compared iterators stemming from different
collections. This isn't defined behavior though :/.

The goal was to determine if a node should add its
children again when moving to the next node in the
iteration. Instead it suffices to keep track of which
nodes have already added their children after adding
them to the ordering.