Hack to make liveness work on Debian testing
A proper solution would be to make the liveness more robust, and in
particular the state equality mechanism. To that extend, we should
 - change the code so that it computes a distance between states
   instead of returning a simple boolean "it's different". That way,
   we could know which states desserve a deeper debugging to
   understand why equality detection is broken, simply by picking the
   closest pair of states, that ought to be equal.

 - Write a bunch of unit tests where we do 2 controled states and
   compare their detected difference.

Maybe, we should trash all this code that is aging (no Dwarf5 support)
and is very complex anyway. A simpler way to get the same effects
could be to go through the compiler, and have clang generate a binary
that is fully introspectable. It sounds like a long and burdensome way
to go, however.

Hopefuly someone will do a library doing it in the future, so that we
just need to use their work.