Remove the stateful model-checking from the archive. It's not working anymore
It's a real pity to remove all this great code, as it was doing very
audacious things, but I need to do so to move forward. This code is
very fragile and badly tested, making it very difficult to build upon.

The next time this is implemented (if any), I guess that a
compiler-approach (e.g. a clang plugin providing a full introspection
over the compiled program) would be more robust than a fully runtime
approach tricking with the debug information to get the info. That's
too difficult and fragile.

Again, I'm deeply sorry I had to remove this code, as it requires me
to admit that I failed to maintain this great code over the years.
But facts are stubborn.