Implement pthread_cond in sthread -- too bad it's TODO in MC
I now have to implement the condition variables in the MC world
(adding the transitions and the dependence theorems), but I first have
to implement the asynchronous version of CV::wait(), as the model
checker can only deal with persistent transitions. That's not a small
assignment, and it will have to wait for tomorrow.

Until then, the pthread_cond support in sthread is somewhat useless.