Fix SMPI mpich test that was broken stupidly 2 months ago
I used THROW_IMPOSSIBLE in kernel space on a code path that is
actually executed. This exception was catched and misinterpreted
somehow. I guess I should have used DIE_IMPOSSIBLE instead, or we
should ensure that IMPOSSIBLE exceptions are not captured in kernel
space but properly kill the simulation.

It's shameful that such a simple fix took 2 months to happen. Sorry
about that.