MC: Prepare to have rollback using intermediate forked applications
The idea is to sometimes fork the application before taking a
transition so that we can restore this state very quickly in the
future. Of course, that's very memory hungry but this is meant to be a
rare event, and it's subject to future optimizations (to remove some
forks when they become useless).

This induced some refactoring to dedupplicate code between DFSExplorer
and OutOfOrderExplorer. That code moved to the ancestor Explorer.