Free the memory of canceled activities ASAP
Each actor has a set of activities it needs to destroy when it gets
killed, but canceled activities were never removed from this list. As
a result, the refcount of canceled activities never reached 0 until
the termination of the actors.

In very cancel-intensive simulations, this could result in a
inordinate memory usage, so removing canceled activities of actors'
lists is better. This makes the code somewhat less tolerent to memory
errors from the user side, as activities are removed very soon. But
this should still be better.